poc/requirements.txt
pageindex>=0.2
openai>=1.0
python-dotenv>=1.0
